fix: address AI review round #3 on the compose autocomplete backport EXO-89067
- sendEmail refuses to send while any recipient row still holds text the field
  rejected, not just when To has no chip at all. "alice@x.com, bad-address"
  previously sent to alice alone and destroyed the warning with the drawer.
  Cc and Bcc now report their pending term too, so leftover text there is no
  longer dropped without a trace; they still never gate Send on their own.
- ADDRESS_PATTERN excludes "." from the domain labels, removing the ambiguity
  that made a near-miss backtrack over every split point. Measured on a dotted
  tail failing at the anchor: 4000 labels went from 28.8ms to 0.11ms, and the
  old form grew 4x per doubling. One deliberate tightening: "a@b..c", an empty
  domain label, is now rejected where it used to pass.
- open() split into drawerTitle / seedRecipients / buildForwardedBody (plus
  quotedRecipients, which de-duplicates the To and Cc header rendering). This
  is the Sonar CRITICAL that was the only condition failing the quality gate.
- Nits: .filter(Boolean) twice, and dropped the redundant (long) cast on
  Stream.limit. Left java:S135 alone -- the three continues are guard clauses
  over a loop that would otherwise nest three deep, as the reviewer suggested.

+++ email-connector-services/src/main/java/org/exoplatform/emailConnector/service/EmailContactService.java
@@ -233,7 +233,7 @@ public class EmailContactService {
     if (StringUtils.isNotBlank(query) && byAddress.size() < size) {
       appendDirectoryMatches(username, query.trim(), size, byAddress);
     }
-    return byAddress.values().stream().limit((long) size).toList();
+    return byAddress.values().stream().limit(size).toList();
   }
 
   /**
